pyproject.toml
[tool.poetry]
name = "4xidraw"
version = "0.1.0"
description = ""
authors = ["Andrew de Quincey <adq@lidskialf.net>"]

[tool.poetry.dependencies]
python = "^3.9"
svg-to-gcode = "^1.5.4"
pyserial = "^3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
